Вынес feature-recipe-detail в отдельный модуль
