+++ app/src/main/java/com/example/otusapp/StartersModule.kt
@@ -1,6 +1,5 @@
 package com.example.otusapp
 
-import com.example.otusapp.recipe.RecipeDetailStarterImpl
 import com.github.belyakovleonid.core.starters.RecipeDetailStarter
 import com.github.belyakovleonid.core.starters.RecipeListStarter
 import com.github.belyakovleonid.feature_recipe_list.starter.RecipeListStarterImpl
@@ -11,7 +10,7 @@ import dagger.Module
 interface StartersModule {
 
     @Binds
-    fun bindRecipeDetailStarter(starter: RecipeDetailStarterImpl): RecipeDetailStarter
+    fun bindRecipeDetailStarter(starter: com.github.belyakovleonid.feature_recipe_detail.starter.RecipeDetailStarterImpl): RecipeDetailStarter
 
     @Binds
     fun bindRecipeListStarter(starter: RecipeListStarterImpl): RecipeListStarter
