--- app/src/main/java/com/example/otusapp/recipe/detail/data/RecipeDetailRepositoryImpl.kt
@@ -1,20 +1,0 @@
-package com.example.otusapp.recipe.detail.data
-
-import com.example.otusapp.recipe.detail.data.remote.RecipeDetailApi
-import com.example.otusapp.recipe.detail.data.remote.model.RecipeDetailDto
-import com.example.otusapp.recipe.detail.data.remote.model.toDomain
-import com.example.otusapp.recipe.detail.domain.RecipeDetailRepository
-import com.example.otusapp.recipe.detail.domain.model.RecipeDetail
-import com.github.belyakovleonid.core_network_api.model.Result
-import com.github.belyakovleonid.core_network_api.model.convertTo
-import javax.inject.Inject
-
-class RecipeDetailRepositoryImpl @Inject constructor(
-    private val recipesApi: RecipeDetailApi
-) : RecipeDetailRepository {
-
-    override suspend fun loadRecipe(id: Long): Result<RecipeDetail> {
-        val result = recipesApi.loadRecipeDetail(id)
-        return result.convertTo(RecipeDetailDto::toDomain)
-    }
-}
