--- app/src/main/java/com/example/otusapp/recipe/detail/domain/model/RecipeDetail.kt
@@ -1,7 +1,0 @@
-package com.example.otusapp.recipe.detail.domain.model
-
-data class RecipeDetail(
-    val id: Long,
-    val name: String,
-    val subtitle: String
-)
