+++ feature-recipe-detail/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.github.belyakovleonid.feature_recipe_detail" />
