+++ feature-recipe-detail/src/main/java/com/github/belyakovleonid/feature_recipe_detail/domain/RecipeDetailInteractor.kt
@@ -1,0 +1,14 @@
+package com.github.belyakovleonid.feature_recipe_detail.domain
+
+import com.github.belyakovleonid.core_network_api.model.Result
+import com.github.belyakovleonid.feature_recipe_detail.domain.model.RecipeDetail
+import javax.inject.Inject
+
+class RecipeDetailInteractor @Inject constructor(
+    private val recipesRepository: RecipeDetailRepository
+) {
+
+    suspend fun loadRecipe(id: Long): Result<RecipeDetail> {
+        return recipesRepository.loadRecipe(id)
+    }
+}
