+++ feature-recipe-detail/src/main/java/com/github/belyakovleonid/feature_recipe_detail/domain/RecipeDetailRepository.kt
@@ -1,0 +1,8 @@
+package com.github.belyakovleonid.feature_recipe_detail.domain
+
+import com.github.belyakovleonid.core_network_api.model.Result
+import com.github.belyakovleonid.feature_recipe_detail.domain.model.RecipeDetail
+
+interface RecipeDetailRepository {
+    suspend fun loadRecipe(id: Long): Result<RecipeDetail>
+}
