+++ feature-recipe-detail/src/main/java/com/github/belyakovleonid/feature_recipe_detail/domain/model/RecipeDetail.kt
@@ -1,0 +1,7 @@
+package com.github.belyakovleonid.feature_recipe_detail.domain.model
+
+data class RecipeDetail(
+    val id: Long,
+    val name: String,
+    val subtitle: String
+)
