+++ feature-recipe-detail/src/main/res/layout/f_recipe_detail.xml
